# docs/api/openapi.json
{
 "openapi": "3.0.0",
 "paths": {
...
  "/users": {
   "get": {
    "operationId": "UsersController_list",
    "parameters": [
     {
      "name": "page",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "limit",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Users"
    ]
   }
  },
...
  "/subscriptions": {
   "get": {
    "operationId": "SubscriptionsController_list",
    "parameters": [],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Subscriptions"
    ]
   },
   "post": {
    "operationId": "SubscriptionsController_create",
    "parameters": [],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateSubscriptionDto"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": ""
     }
    },
    "tags": [
     "Subscriptions"
    ]
   }
  },
...
  "/admin/users": {
   "get": {
    "operationId": "AdminController_listUsers",
    "parameters": [
     {
      "name": "page",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "limit",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Admin"
    ]
   }
  },
  "/admin/users/{id}/status": {
   "patch": {
    "operationId": "AdminController_updateUserStatus",
    "parameters": [
     {
      "name": "id",
      "required": true,
      "in": "path",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "status",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Admin"
    ]
   }
  },
  "/admin/subscriptions": {
   "get": {
    "operationId": "AdminController_listSubscriptions",
    "parameters": [
     {
      "name": "page",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "limit",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Admin"
    ]
   }
  },
  "/admin/commissions": {
   "get": {
    "operationId": "AdminController_listCommissions",
    "parameters": [
     {
      "name": "page",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "limit",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Admin"
    ]
   }
  },
...
  "/admin/payouts": {
   "get": {
    "operationId": "AdminController_listPayouts",
    "parameters": [
     {
      "name": "page",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "limit",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Admin"
    ]
   }
  },
...
  "/admin/licenses": {
   "get": {
    "operationId": "AdminController_listLicenses",
    "parameters": [
     {
      "name": "page",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "limit",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Admin"
    ]
   }
  },
  "/admin/devices": {
   "get": {
    "operationId": "AdminController_listDevices",
    "parameters": [
     {
      "name": "page",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "limit",
      "required": true,
      "in": "query",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": ""
     }
    },
    "tags": [
     "Admin"
    ]
   }
  },
...
 },
 "info": {
  "title": "Predict-A-Trade Control Plane API",
  "description": "XAUUSD Intelligence Platform \u2014 SaaS Control Plane v1.0.0",
  "version": "1.0.0",
  "contact": {}
 },
...
 "servers": [],
 "components": {
  "securitySchemes": {
   "bearer": {
    "scheme": "bearer",
    "bearerFormat": "JWT",
    "type": "http"
   }
  },
  "schemas": {
...
   "CreateSubscriptionDto": {
    "type": "object",
    "properties": {}
   },
...
  }
 }
}